requirements.txt
beautifulsoup4==4.8.2
lxml==4.5.0
pytube3==9.6.4
selenium==3.141.0
